+++ market/src/main/java/university/market/chat/room/service/ChatService.java
@@ -1,8 +1,9 @@
-package university.market.chat.service;
+package university.market.chat.room.service;
 
 import java.util.List;
-import university.market.chat.domain.ChatVO;
-import university.market.chat.service.dto.ChatCreateRequest;
+import university.market.chat.room.domain.ChatMemberVO;
+import university.market.chat.room.domain.ChatVO;
+import university.market.chat.room.service.dto.ChatCreateRequest;
 import university.market.member.domain.MemberVO;
 
 public interface ChatService {
@@ -23,4 +24,6 @@ public interface ChatService {
     void removeMyself(Long chatId);
 
     void updateChat(Long chatId, String title);
+
+    ChatMemberVO getChatMember(Long chatId, Long memberId);
 }
