+++ market/src/test/java/university/market/chat/room/service/ChatServiceTest.java
@@ -1,4 +1,4 @@
-package university.market.chat.service;
+package university.market.chat.room.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -15,14 +15,15 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import university.market.base.exception.BaseException;
-import university.market.chat.domain.ChatMemberVO;
-import university.market.chat.domain.ChatVO;
-import university.market.chat.domain.chatauth.ChatAuthType;
-import university.market.chat.exception.ChatException;
-import university.market.chat.exception.ChatExceptionType;
-import university.market.chat.mapper.ChatMapper;
-import university.market.chat.mapper.ChatMemberMapper;
-import university.market.chat.service.dto.ChatCreateRequest;
+import university.market.chat.room.domain.ChatMemberVO;
+import university.market.chat.room.domain.ChatVO;
+import university.market.chat.room.domain.chatauth.ChatAuthType;
+import university.market.chat.room.exception.ChatException;
+import university.market.chat.room.exception.ChatExceptionType;
+import university.market.chat.room.mapper.ChatMapper;
+import university.market.chat.room.mapper.ChatMemberMapper;
+import university.market.chat.room.service.ChatServiceImpl;
+import university.market.chat.room.service.dto.ChatCreateRequest;
 import university.market.helper.fixture.ChatFixture;
 import university.market.helper.fixture.ChatMemberFixture;
 import university.market.helper.fixture.ItemFixture;
